weather/requirements.txt
asgiref==3.8.1
attrs==23.2.0
cattrs==23.2.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
Django==5.0.7
djangorestframework==3.15.2
flatbuffers==24.3.25
future==1.0.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
h3==3.7.7
idna==3.7
numpy==2.0.0
openmeteo_requests==1.2.0
openmeteo_sdk==1.11.11
pandas==2.2.2
platformdirs==4.2.2
psycopg2-binary==2.9.9
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2024.1
ratelim==0.1.6
requests==2.32.3
requests-cache==1.2.1
retry-requests==2.0.0
setuptools==70.3.0
six==1.16.0
sqlparse==0.5.1
timezonefinder==6.5.2
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.2
